Enable fetching of config files by URL
This commit also adds a bunch of stuff that improves error handling,
including making it more verbose and actually reporting it to the user,
as well some emoji-based indicators that display what the app is doing.
I probably should have tried to split these into different commits, but
they were so closely intertwined, I didn't bother. And it's not like
anyone is ever going to need to partially revert this commit.