Get initial defaults from URL query
The tool is intended to be used by embedding it as a script in a
[[html]] block on Wikidot. [[html]] is used instead of a standard iframe
in order to leverage automatic iframe resizing, which Wikidot provides
automatically for [[html]] blocks. This results in URL query parameters
needing to be applied to the URL of the script itself rather than to the
URL of the iframe. To that effect, the query parameter getter extracts
its parameters from the URL of the script.

The implementation of this requires that the script element has an id
'script' so that the function can easily detect it in the document.